Refactor: remove second ParseMessageType; remove gjson dependency
Instead of verifying before parsing just parse the byte array to
message.Message on register and just parse immediately on any other
case.

This also removes the dependency to gjson.
