+++ cmd/example-client/main.go
@@ -76,21 +76,12 @@ func (e *ExampleHandler) On(
 ) (*connection.SendResponse, error) {
 	e.Lock()
 	defer e.Unlock()
-	mt, err := handler.ParseMessageType(msg)
-	if err != nil {
-		// In this example we end the program on a unexpected message so that we
-		// can
-		// reuse it as a smoke test.
-		// However in a production environment you want to either log and ignore
-		// or
-		// just ignore unparseable messages.
-		panic(err)
-	}
-	log.Printf("Got message: %s", mt)
+
 	var infoMSG info.IDInfo
 	if err := json.Unmarshal(msg, &infoMSG); err != nil {
 		log.Fatal().Msgf("Unable to parse %s to info.IDInfo (%s)", msg, err)
 	}
+	mt := infoMSG.MessageType()
 	f, ok := e.do[infoMSG.ID]
 	if !ok {
 		f, ok = e.do[mt.String()]
