+++ go.mod
@@ -13,7 +13,6 @@ require (
 	github.com/pelletier/go-toml v1.9.4
 	github.com/rogpeppe/go-internal v1.8.0 // indirect
 	github.com/rs/zerolog v1.26.0
-	github.com/tidwall/gjson v1.11.0
 	golang.org/x/sys v0.0.0-20211110154304-99a53858aa08 // indirect
 	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
